Move Log to Abnormality/Log
