+++ src/app/components/Abnormality/Log/index.tsx
