+++ src/app/components/Abnormality/Log/log.module.scss
