+++ src/app/components/Abnormality/Log/tabs.module.scss
