+++ src/app/components/Filter/SinnerToggle/index.module.scss
@@ -1,0 +1,34 @@
+.toggle-group {
+  display: flex;
+  flex-flow: wrap row;
+  gap: 0.25rem;
+  place-items: center;
+
+  button {
+    justify-content: center;
+
+    margin: 1.5px;
+    padding: 2px 12px;
+
+    font-size: 16px;
+    font-weight: bold;
+    color: #341408;
+
+    background-color: var(--fg-secondary);
+    border: 2px solid var(--accent);
+
+    transition: 0.15s;
+
+    &:hover {
+      color: var(--fg-primary);
+      background-color: var(--bg-secondary);
+      border: 2px solid var(--border-hover);
+    }
+
+    &[data-state='on'] {
+      color: #ffd57b;
+      background-color: #663c2b7e;
+      border: 2px solid #ffd57b;
+    }
+  }
+}
